+++ src/semaphore.c
@@ -158,7 +158,7 @@ _dispatch_group_create_with_count(uint32_t n)
 	dg->do_targetq = _dispatch_get_default_queue(false);
 	if (n) {
 		os_atomic_store2o(dg, dg_bits,
-				-n * DISPATCH_GROUP_VALUE_INTERVAL, relaxed);
+				(uint32_t)(-n * DISPATCH_GROUP_VALUE_INTERVAL), relaxed);
 		os_atomic_store2o(dg, do_ref_cnt, 1, relaxed); // <rdar://22318411>
 	}
 	return dg;
